Split simplemockwebserver
